Rename io modules to io_ and fix some version conflicts
Numpy 1.16.* is not compatible with sparsity 0.20.* thus we need to fix
the setup.py. When using Scipy<1.0.0 empty column access does not work,
thus the dependency had to be adjusted here as well.
This also renames the io_ modules to avoid issues with pythons
internal module.
