+++ sparsity/dask/__init__.py
@@ -1,5 +1,5 @@
 from .core import SparseFrame
-from .io import from_pandas, read_npz, from_ddf
+from .io_ import from_pandas, read_npz, from_ddf
 from .reshape import one_hot_encode
